\name{CMLS-internal}

%% List of Internal Functions Called
\alias{IsplineBasis}
\alias{gcvfun}
\alias{kcvmae}
\alias{kcvmse}
\alias{list2bdiag}
\alias{MsplineBasis}
\alias{PseudoInverse}
\alias{rsinvsqrt}
\alias{solveQP}
\alias{solveQPerror}

\title{Internal 'CMLS' Functions}
\description{
  Internal functions for 'CMLS' package.
}
\details{
  These functions are not to be called by the user.
}
\keyword{ internal }
